+++ mobile/test/ui/pages/auth/post_login_destination_test.dart
@@ -160,8 +160,13 @@ class _FakeAuthRepository implements AuthRepository {
       throw UnimplementedError();
 
   @override
-  AsyncResult<OperationalAuthState> login(LoginRequestDto dto) async =>
+  AsyncResult<AuthState> login(LoginRequestDto dto) async =>
       throw UnimplementedError();
+
+  @override
+  AsyncResult<OperationalAuthState> certifyInstallation(
+    String transactionPassword,
+  ) async => throw UnimplementedError();
 
   @override
   AsyncResult<Unit> logout() async => throw UnimplementedError();