+++ mobile/test/ui/pages/auth/transaction_password/viewmodel/transaction_password_viewmodel_test.dart
@@ -90,4 +90,11 @@ class _FakeTransactionPasswordRepository
   ) {
     throw UnimplementedError();
   }
+
+  @override
+  AsyncResult<StepUpAuthorizeResponseDto> authorizeInstallationRegistration(
+    String transactionPassword,
+  ) {
+    throw UnimplementedError();
+  }
 }