+++ mobile/test/ui/pages/splash/viewmodel/splash_viewmodel_test.dart
@@ -96,8 +96,13 @@ class _FakeAuthRepository implements AuthRepository {
   }
 
   @override
-  AsyncResult<OperationalAuthState> login(LoginRequestDto dto) async =>
+  AsyncResult<AuthState> login(LoginRequestDto dto) async =>
       throw UnimplementedError();
+
+  @override
+  AsyncResult<OperationalAuthState> certifyInstallation(
+    String transactionPassword,
+  ) async => throw UnimplementedError();
 
   @override
   AsyncResult<Unit> logout() async => throw UnimplementedError();