+++ mobile/test/ui/pages/transaction_password/setup/create_transaction_password_page_test.dart
@@ -70,4 +70,11 @@ class _FakeTransactionPasswordRepository
   ) {
     throw UnimplementedError();
   }
+
+  @override
+  AsyncResult<StepUpAuthorizeResponseDto> authorizeInstallationRegistration(
+    String transactionPassword,
+  ) {
+    throw UnimplementedError();
+  }
 }